change user from subject to behavior subject
Subject will subscribe to all future emits of the subject but behavior will do that also but will also get the last emit. this is important in our data service and other places so that it can see if the user is actively authenticated because maybe that info, especially token info, would have been emitted at that time. now with behaviorsubject it will be.